+++ app/src/main/java/com/example/weather/data/web/model/yandex/NightShort.kt
@@ -17,7 +17,7 @@ data class NightShort(
 	val icon: String? = null,
 
 	@field:SerializedName("cloudness")
-	val cloudness: Int? = null,
+	val cloudness: Float? = null,
 
 	@field:SerializedName("prec_period")
 	val precPeriod: Int? = null,
